Added Ladder and Interactable logic.
Created Ladder prefab with associated logic - needs tweaking, but
essentially does what we need it to do, minus animations and constraint
triggers.

Also added Interactable code - things in the world that are interactable
will have to subclass the Interactable class and specify a cooldown.

When the player hits the Interact button ("E"), all Interactables
currently in range will be triggered.